change int for decimal at sql file

+++ db/script_db.sql
@@ -12,10 +12,10 @@ id int auto_increment primary key,
 currency_id int not null,
 frequency int not null,
 reg_date TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-open_value int not null,
-low_value int not null,
-high_value int not null,
-close_value int not null,
+open_value decimal(9,8) not null,
+low_value decimal(9,8) not null,
+high_value decimal(9,8) not null,
+close_value decimal(9,8) not null,
 foreign key (currency_id) references currencies(id)
 );
